+++ db4e/Modules/XMRig.py
@@ -22,7 +22,8 @@ from db4e.Constants.DDef import DDef
 from db4e.Constants.DPlaceholder import DPlaceholder
 
 from db4e.Modules.Components import (
-    ConfigFile, Enabled, Instance, Local, LogFile, NumThreads, Parent, Version)
+    ConfigFile, Enabled, Instance, Local, LogFile, NumThreads, Parent, Version,
+    MaxLogFiles, MaxLogSize, LogRotateConfig)
 
 
 class XMRig(SoftwareSystem):
@@ -36,6 +37,9 @@ class XMRig(SoftwareSystem):
         self.add_component(DField.ENABLED, Enabled())
         self.add_component(DField.INSTANCE, Instance())
         self.add_component(DField.LOG_FILE, LogFile())
+        self.add_component(DField.LOG_ROTATE_CONFIG, LogRotateConfig())
+        self.add_component(DField.MAX_LOG_FILES, MaxLogFiles())
+        self.add_component(DField.MAX_LOG_SIZE, MaxLogSize())
         self.add_component(DField.REMOTE, Local())
         self.add_component(DField.NUM_THREADS, NumThreads())
         self.add_component(DField.VERSION, Version())
@@ -45,6 +49,9 @@ class XMRig(SoftwareSystem):
         self.enabled = self.components[DField.ENABLED]
         self.instance = self.components[DField.INSTANCE]
         self.log_file = self.components[DField.LOG_FILE]
+        self.logrotate_config = self.components[DField.LOG_ROTATE_CONFIG]
+        self.max_log_files = self.components[DField.MAX_LOG_FILES]
+        self.max_log_size = self.components[DField.MAX_LOG_SIZE]
         self.num_threads = self.components[DField.NUM_THREADS]
         self.parent = self.components[DField.PARENT]
         self.version = self.components[DField.VERSION]
@@ -89,6 +96,32 @@ class XMRig(SoftwareSystem):
         with open(fq_config, 'w') as f:
             f.write(final_config)
         self.config_file(fq_config)
+
+
+    def gen_logrotate_config(self, tmpl_file: str, vendor_dir: str, db4e_group:str):
+        # Logrotate configuration file
+        fq_config = os.path.join(
+            vendor_dir, DElem.DB4E, DDef.LOG_ROTATE, DElem.XMRIG + "-" + \
+                self.instance() + DDef.CONF_SUFFIX)
+        
+        # Populate the config template placeholders
+        placeholders = {
+            DPlaceholder.VENDOR_DIR: vendor_dir,
+            DPlaceholder.INSTANCE: self.instance(),
+            DPlaceholder.MAX_LOG_FILES: self.max_log_files(),
+            DPlaceholder.MAX_LOG_SIZE: self.max_log_size(),
+            DPlaceholder.DB4E_GROUP: db4e_group,
+        }
+        with open(tmpl_file, 'r') as f:
+            config_contents = f.read()
+            final_config = config_contents
+            for key, val in placeholders.items():
+                final_config = final_config.replace(f'[[{key}]]', str(val))
+
+        # Write the config to file
+        with open(fq_config, 'w') as f:
+            f.write(final_config)
+        self.logrotate_config(fq_config)
 
 
     def hashrate(self, hashrate=None):